added mappings for period types and also added templates (add, edit and list) also added to class Bootstrap static period types and tarifs

+++ src/main/java/kg/inaiproject/booking/entities/Tarif.java
@@ -23,5 +23,5 @@ public class Tarif {
     String name;
 
     @Column(name = "price")
-    BigDecimal price;
+    Double price;
 }
